Move Article service to `service`
